Allow the mail transporter to be configured from the db
This commit also adds a new siteConfig object which is used to
refactor all of the mail configuration in mail.js.

siteConfig is filled with the site data through the site middleware,
but loads the current config data (from the config object) before
merging the data from the database. This allows us to specify a SMTP
or sendgrid server in the database config instead of the local.json
file.

Currently the siteConfig is only used in the mail configuration, but
it can be expanded to be used in other places as well.

# Conflicts:
#	package-lock.json
#	src/lib/mail.js